[micrOS] WebEngine improvements, chunked encoding
Allow resources to be sent in chunks depending on
a predefined size threshold. This way, boards with
lower SRAM sizes can also serve the Web UI. This commit
decreases the minimum RAM to 20 KB to serve the Web UI
based on experiments with an ESP32-C3.

Introduce HTTP 503 (server busy) and 500 (server error)
to allow more verbose load-module behavior.

Change the default error response code (404) to 400
to be less ambiguous when a generic exception is
thrown not due to a missing resource.

Remove only leading '/' characters when loading
web resources so that arbitrarily deep resource paths
are handled.